Methods list

ImagePhantoms.AbstractShape

ImagePhantoms.AbstractShape2

ImagePhantoms.AbstractShape3

ImagePhantoms.Ellipse

ImagePhantoms.Ellipse

ImagePhantoms.Ellipse

ImagePhantoms.EllipsePhantomVersion

ImagePhantoms.Gauss2

ImagePhantoms.Gauss2

ImagePhantoms.Object

ImagePhantoms.Object

ImagePhantoms.Object

ImagePhantoms.Object

ImagePhantoms.Object

ImagePhantoms.Object

ImagePhantoms.Object2d

ImagePhantoms.Object3d

ImagePhantoms.Rect

ImagePhantoms.Rect

ImagePhantoms.SheppLogan

ImagePhantoms.SheppLoganToft

ImagePhantoms.Triangle

ImagePhantoms.Triangle

Base.:*

Base.show

ImagePhantoms.Circle

ImagePhantoms.Square

ImagePhantoms._interval

ImagePhantoms.coords

ImagePhantoms.disk_phantom_params

ImagePhantoms.ellipse_parameters

ImagePhantoms.ellipse_parameters

ImagePhantoms.focus_chart

ImagePhantoms.fwhm2spread

ImagePhantoms.jinc

ImagePhantoms.mri_smap_basis

ImagePhantoms.mri_smap_fit

ImagePhantoms.mri_spectra

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.phantom

ImagePhantoms.radon

ImagePhantoms.radon

ImagePhantoms.radon

ImagePhantoms.radon

ImagePhantoms.radon

ImagePhantoms.radon

ImagePhantoms.radon_ellipse

ImagePhantoms.radon_gauss2

ImagePhantoms.radon_rect

ImagePhantoms.radon_tri

ImagePhantoms.radon_tri

ImagePhantoms.rotate

ImagePhantoms.rotate

ImagePhantoms.scale

ImagePhantoms.shepp_logan

ImagePhantoms.shepp_logan_values

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.spectrum

ImagePhantoms.translate

ImagePhantoms.trapezoid

Methods usage

ImagePhantoms.AbstractShape — Type

AbstractShape

Generic shape type for ImagePhantoms.

ImagePhantoms.AbstractShape2 — Type

AbstractShape2 <: AbstractShape

Generic 2D shape type (with subtypes Ellipse...)

ImagePhantoms.AbstractShape3 — Type

AbstractShape2 <: AbstractShape

Generic 3D shape type (with subtypes Ellipsoid...)

ImagePhantoms.Ellipse — Type

Ellipse(cx, cy, rx=1, ry=rx, ϕ=0, value::Number=1)
Ellipse(center::NTuple{2,RealU}, radii::NTuple{2,RealU}=(1,1), ϕ::RealU=0, v=1)
Ellipse([6-vector])

Construct Ellipse object from parameters.

ImagePhantoms.Ellipse — Type

Ellipse <: AbstractShape2

ImagePhantoms.Ellipse — Method

phantom = Ellipse(n × 6 AbstractMatrix)

Return vector of Ellipse objects, one for each row of input matrix.

ImagePhantoms.EllipsePhantomVersion — Type

EllipsePhantomVersion

Parent type for different versions of ellipse phantoms:

SheppLogan original CT version from Shepp&Logan paper

SheppLoganEmis higher contrast version suitable for emission tomography

SheppLoganBrainWeb integer index version based on brainweb

SheppLoganToft higher contrast version from Toft, 1996

SouthPark for fun

ImagePhantoms.Gauss2 — Type

Gauss2(cx, cy, wx, wy=wx, ϕ=0, value::Number=1)
Gauss2(center::NTuple{2,RealU}, width::NTuple{2,RealU}=(1,1), ϕ::RealU=0, v=1)
Gauss2([6-vector])
Gauss2(w, v=1) (isotropic of width `w`)

Construct Gauss2 object from parameters; here width = FWHM (full-width at half-maximum).

In 1D, the formula is g(x) = exp(-π ((x - cx) / sx)^2) where sx = fwhm2spread(w) = w * sqrt(π / log(16)), which, for cx=0, has 1D FT G(ν) = sx^2 exp(π (sx νx)^2).

ImagePhantoms.Gauss2 — Type

Gauss2 <: AbstractShape2

ImagePhantoms.Object — Type

Object{S, D, V, ...}(center, width, angle, value, param) <: AbstractObject

General container for 2D and 3D objects for defining image phantoms.

center::NTuple{D,C} coordinates of "center" of this object

width::NTuple{W,C} "width" along axis; usually W=D or W=1 (e.g., FWHM for Gauss, radii for Ellipse, radius of Circle)

angle::NTuple{D-1,A} angle of x' axis relative to x axis, in radians (or with units)

value::V "intensity" value for this object

param optional additional shape parameters (often nothing)

Example

julia> Object(Ellipse(), (0,0), (1,2), 0.0, 1//2, nothing)
Object2d{Ellipse, Rational{Int64}, 2, Int64, Float64, Nothing} (S, D, V, ...)
 shape::Ellipse Ellipse()
 center::NTuple{2,Int64} (0, 0)
 width::NTuple{2,Int64} (1, 2)
 angle::Tuple{Float64} (0.0,)
 value::Rational{Int64} 1//2
 param::Nothing nothing

ImagePhantoms.Object — Method

ob = Object(shape ; center=(0,0), width=(1,1), angle=0, value=1, param=nothing)

2D object constructor using keywords.

ImagePhantoms.Object — Method

ob = Object(shape ; center=(0,0,0), width=(1,1,1), angle=(0,0), value=1, param=nothing)

3D object constructor using keywords.

ImagePhantoms.Object — Method

ob = Object(shape, center, width, angle=0, value=1, param=nothing)

2D object constructor.

ImagePhantoms.Object — Method

ob = Object(shape, center, width, angle, value, param)

General outer object constructor where angle is a tuple, including 3D case.

ImagePhantoms.Object — Method

Object(ob::Object ; center, width, angle, value, param)

Make a copy of Object ob, optionally modifying some values.

ImagePhantoms.Object2d — Type

Object2d = Object{S,2} where S <: AbstractObject

For 2D objects

ImagePhantoms.Object3d — Type

Object3d = Object{S,3} where S <: AbstractObject

For 3D objects

ImagePhantoms.Rect — Type

Rect(cx, cy, wx=1, wy=wx, ϕ=0, value::Number=1)
Rect(center::NTuple{2,RealU}, width::NTuple{2,RealU}=(1,1), ϕ::RealU=0, v=1)
Rect([6-vector])

Construct Rect object from parameters; here width is the full-width.

ImagePhantoms.Rect — Type

Rect <: AbstractShape2

ImagePhantoms.SheppLogan — Type

SheppLogan

Original version from : Larry A Shepp, Benjamin F Logan, "The Fourier reconstruction of a head section," IEEE Transactions on Nuclear Science, 21(3):21-42, June 1974. doi

Also in Kak and Slaney 1988 text, p. 255. doi

ImagePhantoms.SheppLoganToft — Type

SheppLoganToft

Toft, Peter Aundal & Sørensen, John Aasted "The Radon transform-theory and implementation," Technical University of Denmark (DTU), 1996. Page 201. https://files.openpdfs.org/1ra51GP6gJO.pdf

ImagePhantoms.Triangle — Type

Triangle(cx, cy, wx=1, wy=wx, ϕ=0, value::Number=1, param::Real=0.5)
Triangle(center::NTuple{2,RealU}, width::NTuple{2,RealU}=(1,1), ϕ::RealU=0, v=1, param=0.5)
Triangle([6-vector] or [7-vector])

Construct Triangle object from parameters. In the typical case where param=0.5 and width[1] == width[2], this is an equilateral triangle with base width[1] centered along the x axis.

ImagePhantoms.Triangle — Type

Triangle{T} <: AbstractShape2

By default an equilateral triangle pointing upward with the "center" in the middle of the base, for the default parameter = 0.5.

For parameter p the object is a triangle whose base is along the x-axis going from (-p,0) to (1-p,0)` and with height=sqrt(3)/2.

Most methods currently support only the case p=0.5.

Base.:* — Method

(*)(ob::Object, x::Number)
(*)(x::Number, ob::Object)

Scale object value by x.

Base.show — Method

show(io::IO, ::MIME"text/plain", ob::Object)

ImagePhantoms.Circle — Function

Circle(x,y,r,v=1) (circle of radius `r` centered at `(x,y)`)
Circle((x,y), r=1, v=1) ditto
Circle([4-vector]) ditto
Circle(r, v=1) centered at origin

Construct circle objects as special cases of Ellipse objects.

ImagePhantoms.Square — Function

Square(x,y,w,v=1) (square of width `w` centered at `(x,y)`)
Square((x,y), w=1, v=1) ditto
Square([4-vector]) ditto
Square(w, v=1) centered at origin

Construct Square objects as special cases of Rect objects.

ImagePhantoms._interval — Method

(lower, upper) = _interval(a, b)

Determine the interval [lower, upper] corresponding to the set {x : b ≤ a x} where a is unitless but b and x have same units.

ImagePhantoms.coords — Method

coords(object::Object2d, x::RealU, y::RealU)

Put coordinates (x,y) in canonical axes associated with object.

ImagePhantoms.disk_phantom_params — Method

params = disk_phantom_params( ; ...)

Generate ndisk × 6 ellipse phantom parameters for a head-sized disk plus many disks within it, designed so that the disks have some minimum separation minsep to avoid overlap and to simplify patch-based model fitting.

Options

fov::Real = 240 image field of view in mm

rhead::Function = () -> 100 background radius for "head" [mm]

muhead::Function = () -> 1000 "μ" (intensity) value for background head disk

rdisk::Function = () -> 10 + 10 * rand() random disk radii [10,20]

mudisk::Function = () -> 100 + 200 * rand() "μ" values for disks [100,300]

ndisk::Function = () -> 10 # of random disks

minsep::Real = 8 minimum disk separation in mm

maxtry::Int = 500 give up on adding more disks if this is reached

warn::Bool = false warn if maxtry reached?

seed::Int = 0 if nonzero then use this seed

The function options can be replaced with rand() for other Distributions.

ImagePhantoms.ellipse_parameters — Function

params = ellipse_parameters(case::Symbol; fovs::NTuple{2}, u::Tuple, disjoint::Bool)

By default the first four columns are unitless "fractions of field of view", so columns 1,3 are scaled by xfov and columns 2,4 are scaled by yfov, where (xfov, yfov) = fovs. The optional 3-tuple u specifies scaling and/or units:

columns 1-4 (center, radii) are scaled by u[1] (e.g., mm),

column 5 (angle) is scaled by u[2] (e.g., 1 or °),

column 6 (value) is scaled by u[3] (e.g., 1/cm) for an attenuation map.

If disjoint==true then the middle ellipse positions are adjusted to avoid overlap.

ImagePhantoms.ellipse_parameters — Method

params = ellipse_parameters(SouthPark() ; fovs::NTuple{2,Number} = (100,100))

Ellipse parameters for "South Park" phantom.

ImagePhantoms.focus_chart — Method

focus_chart( ; radius=1, nspoke=30, value=1)

Generate nspoke Triangle phantom parameters for a focus chart. Returns Vector{Object2d{Triangle} for passing to phantom.

Options

radius::RealU = 1 radius of phantom

nspoke::Int = 60 # of spokes

value::Number = 1 alternate between 0 and this value

ImagePhantoms.fwhm2spread — Method

s = fwhm2spread(w)

Convert FWHM w to equivalent Gaussian spread s for $\exp(-π (x/s)^2)$. exp(-π (fwhm/2/s)^2) = 1/2 means fwhm/2/s) = sqrt(log(2)/π) 2s/fwhm = sqrt(π/log(2)) 2s = fwhm * sqrt(π/log(2)) s = fwhm * sqrt(π / log(16))

ImagePhantoms.jinc — Method

jinc(x)

Return jinc(x) = J1(π*x)/(2x), where J1 is a Bessel function of the first kind.

The argument x must be unitless.

Return type is promote_type(typeof(x), Float32).

ImagePhantoms.mri_smap_basis — Method

mrismapbasis(mask ; kmax, kt, ki)

Construct Fourier basis for representing MRI sensitivity maps in terms of separable complex exponential signals, products of of the form basis = (k,N) -> exp.(2im * π * nfun(N) * kfun(k,N)). The default is nfun(N) = -(N÷2):(N÷2)-1 which is suitable for even N only. The default is kfun(k,N) = k / 2N, which is DCT-II like frequencies, leading to better boundary behavior than the DFT frequencies k/N.

Input

mask::AbstractArray{Bool,D} binary support mask for region to reconstruct

Option

kmax::Int = 5 default frequency index -kmax:kmax in all dimensions

kfun::Function = (k,N) -> k / (2N) # DCT-II frequency

deltas::NTuple{D,<:Number} = ones(D) pixel sizes

(For additional options kmaxs, kt, ki, see code.)

Output

(; B, ν) where B is basis matrix of size count(mask) × nk where typically nk = (2*kmax+1)^D and ν is nk frequency tuples; each tuple has form ν = kfun.(Tuple(k), size(mask)) ./ deltas.

ImagePhantoms.mri_smap_fit — Method

mri_smap_fit(smaps, embed ; mask, kwargs...)

Fit MRI sensitivity maps smaps using mri_smap_basis(mask ; kwargs...). Caller provides ImageGeoms.embed or equivalent.

Return named tuple (B, ν, coefs, nrmse, smaps):

(B, ν) from mri_smap_basis

coefs::Vector : [ncoil] each of length nk

nrmse::Real : smaps vs smaps_fit

smaps::Vector{Array{D}} : smaps_fit

ImagePhantoms.mri_spectra — Method

mri_spectra(fx, fy, oa::Array{<:Object2d}, fit::NamedTuple)

Version of spectrum suitable for parallel MRI with sensitivity maps that were fit previously using mri_smap_fit. Returns a Vector of ncoil kspace data Vectors of dimension length(fx)

ImagePhantoms.phantom — Method

image = phantom(x, y, oa::Array{<:Object2d}, oversample::Int; T)

Return a digital image of the phantom sampled at (x,y) locations, with over-sampling factor oversample and element type T.

ImagePhantoms.phantom — Method

image = phantom(x, y, oa::Array{<:Object2d})

Return a digital image of the phantom sampled at (x,y) locations.

ImagePhantoms.phantom — Method

image = phantom(oa::Array{<:Object2d})::Function

Return function image(x,y) that user can sample at any (x,y) locations to make a phantom image.

ImagePhantoms.phantom — Method

phantom(ob::Object2d{Ellipse})

Returns function of (x,y) for making image.

ImagePhantoms.phantom — Method

phantom(ob::Object2d{Gauss2})

Returns function of (x,y) for making image.

ImagePhantoms.phantom — Method

phantom(ob::Object2d{Rect})

Returns function of (x,y) for making image.

ImagePhantoms.phantom — Method

phantom(ob::Object2d{Triangle})

Returns function of (x,y) for making image.

ImagePhantoms.radon — Method

sino = radon(r, ϕ, oa::Array{<:Object2d})

Return parallel-beam 2D sinogram sino sampled at given (r,ϕ) locations.

ImagePhantoms.radon — Method

sino = radon(oa::Array{<:Object2d})::Function

Return function sino(r,ϕ) that user can sample at any (r,ϕ) locations to make a phantom 2D sinogram.

The coordinate system used here is such that ϕ=0 corresponds to line integrals along the $y$ axis for an object $f(x,y)$. Then as ϕ increases, the line integrals rotate counter-clockwise.

ImagePhantoms.radon — Method

radon(ob::Object2d{Ellipse})

Returns function of (r,ϕ) for making a sinogram.

ImagePhantoms.radon — Method

radon(ob::Object2d{Gauss2})

Returns function of (r,ϕ) for making a sinogram.

ImagePhantoms.radon — Method

radon(ob::Object2d{Rect})

Returns function of (r,ϕ) for making a sinogram.

ImagePhantoms.radon — Method

radon(ob::Object2d{Triangle})

Returns function of (r,ϕ) for making a sinogram.

ImagePhantoms.radon_ellipse — Method

radon_ellipse(r, ϕ, cx, cy, rx, ry, θ)

Radon transform of ellipse at point (r,ϕ).

ImagePhantoms.radon_gauss2 — Method

radon_gauss2(r, ϕ, cx, cy, wx, wy, θ)

Radon transform at (r,ϕ) of 2D Gaussian.

ImagePhantoms.radon_rect — Method

radon_rect(r, ϕ, cx, cy, wx, wy, θ)

Radon transform at (r,ϕ) of rectangle (involves a trapezoid).

ImagePhantoms.radon_tri — Method

radon_tri(r, sinϕ, cosϕ)

For a line integral at radial position r and angle ϕ, the locus of points along the line is {(r cos(ϕ), r sin(ϕ)) + τ (-sin(ϕ), cos(ϕ)) : τ ∈ ℝ}. This function treats the equilateral triangle with base [-1/2,1/2], pointing upwards, as the intersection of three half planes:

H0 = {(x,y) : y ≥ 0}

H1 = y ≤ √3 (1/2 - x)

H2 = y ≤ √3 (1/2 + x).

Find the τ values where the line locus lies in each half planes, then take the length of the intersection of those three intervals.

For example, for H1 we have r sin(ϕ) + τ cos(ϕ) ≤ √3 (1/2 - (r cos(ϕ) - τ sin(ϕ))) or equivalently r (sin(ϕ) + √3 cos(ϕ)) - √3/2 ≤ τ (√3 sin(ϕ) - cos(ϕ)) which is a set the form b1 ≤ a1 τ, corresponding to some interval (l1,u1). Similarly for H0 and H2.

This approach might not be the most efficient, but it is simple.

See Peter Aundal Toft, "The Radon transform - theory and implementation", 1996 https://orbit.dtu.dk/en/publications/the-radon-transform-theory-and-implementation for a different approach to finding the Radon transform of a triangle.

ImagePhantoms.radon_tri — Method

radon_tri(r, ϕ, cx, cy, wx, wy, θ, p)

Radon transform at (r,ϕ) of triangle.

ImagePhantoms.rotate — Method

rotate(ob::Object2d, θ::RealU)

In-plane rotation of a 2D object.

ImagePhantoms.rotate — Method

rotate(ob::Object3d, (α,β))
rotate(ob::Object3d, α, β=0)

Rotation of a 3D object.

ImagePhantoms.scale — Method

scale(ob::Object, factor::RealU)
scale(ob::Object, factor::NTuple{W,RealU})

Scale the width(s) by factor.

ImagePhantoms.shepp_logan — Function

image = shepp_logan(M, [N,], case, options...)

Convenience method for generating M×N samples of Shepp-Logan phantoms.

In

M::Int : horizontal size

N::Int : vertical size, defaults to M

case::EllipsePhantomVersion = SheppLogan()

Options

oversample::Int = 3 (usually)

yflip::Bool = true (reverse y samples for convenience.)

kwargs... remaining options passed to ellipse_parameters for parameters.

Out

image : M × N matrix

The default here is 3× over-sampling along both axes (9 samples per pixel), except for the SheppLoganBrainWeb phantom that consists of integer indices.

ImagePhantoms.shepp_logan_values — Method

values = shepp_logan_values(::EllipsePhantomVersion)

Return 10 Shepp-Logan ellipse amplitudes for various versions.

ImagePhantoms.spectrum — Method

kspace = spectrum(fx, fy, oa::Array{<:Object2d})::Function

Return k-space matrix kspace sampled at given (fx,fy) locations.

ImagePhantoms.spectrum — Method

kspace = spectrum(oa::Array{<:Object2d})::Function

Return function kspace(fx,fy) that user can sample at any (fx,fy) locations to make phantom 2D k-space data.

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d, coefs::AbstractVector, f::)

Version of spectrum(ob) suitable for parallel MRI with sensitivity maps that were fit previously using mri_smap_fit for a single coil with fit coefficients coefs and frequencies f (array of tuples).

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d, fit::NamedTuple, coil::Int)

Version of spectrum(ob) suitable for parallel MRI with sensitivity maps that were fit previously using mri_smap_fit.

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d{Ellipse})

Returns function of $(f_x,f_y)$ for the spectrum (2D Fourier transform).

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d{Gauss2})

Returns function of $(f_x,f_y)$ for the spectrum (2D Fourier transform).

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d{Rect})

Returns function of $(f_x,f_y)$ for the spectrum (2D Fourier transform).

ImagePhantoms.spectrum — Method

spectrum(ob::Object2d{Triangle})

Returns function of $(f_x,f_y)$ for the spectrum (2D Fourier transform).

ImagePhantoms.translate — Method

translate(ob::Object, shift::NTuple{D,RealU})
translate(ob::Object2d, xshift, yshift)
translate(ob::Object3d, xshift, yshift, zshift)

Translate the center coordinates of an object by shift

ImagePhantoms.trapezoid — Method

trapezoid(t::RealU, t1, t2, t3, t4)

Unit-height trapezoid with breakpoints t1, t2, t3, t4.